Split color space details in their own files
