+++ src/UglyToad.PdfPig/Graphics/Colors/ColorSpaces/DeviceGrayColorSpaceDetails.cs
@@ -1,0 +1,74 @@
+﻿namespace UglyToad.PdfPig.Graphics.Colors
+{
+    using System;
+
+    /// <summary>
+    /// A grayscale value is represented by a single number in the range 0.0 to 1.0,
+    /// where 0.0 corresponds to black, 1.0 to white, and intermediate values to different gray levels.
+    /// </summary>
+    public sealed class DeviceGrayColorSpaceDetails : ColorSpaceDetails
+    {
+        /// <summary>
+        /// The single instance of the <see cref="DeviceGrayColorSpaceDetails"/>.
+        /// </summary>
+        public static readonly DeviceGrayColorSpaceDetails Instance = new DeviceGrayColorSpaceDetails();
+
+        /// <inheritdoc/>
+        public override int NumberOfColorComponents => 1;
+
+        /// <inheritdoc/>
+        public override int BaseNumberOfColorComponents => NumberOfColorComponents;
+
+        private DeviceGrayColorSpaceDetails() : base(ColorSpace.DeviceGray)
+        { }
+
+        /// <inheritdoc/>
+        internal override double[] Process(params double[] values)
+        {
+            return values;
+        }
+
+        /// <inheritdoc/>
+        public override IColor GetColor(ReadOnlySpan<double> values)
+        {
+            if (values.Length != NumberOfColorComponents)
+            {
+                throw new ArgumentException($"Invalid number of inputs, expecting {NumberOfColorComponents} but got {values.Length}", nameof(values));
+            }
+
+            double gray = values[0];
+            if (gray == 0)
+            {
+                return GrayColor.Black;
+            }
+
+            if (gray == 1)
+            {
+                return GrayColor.White;
+            }
+
+            return new GrayColor(gray);
+        }
+
+        /// <inheritdoc/>
+        public override IColor GetInitializeColor()
+        {
+            return GrayColor.Black;
+        }
+
+        /// <inheritdoc/>
+        public override void GetRgb(ReadOnlySpan<double> values, out double r, out double g, out double b)
+        {
+            double gray = values[0];
+            r = gray;
+            g = gray;
+            b = gray;
+        }
+
+        /// <inheritdoc/>
+        internal override Span<byte> Transform(Span<byte> decoded)
+        {
+            return decoded;
+        }
+    }
+}
