+++ src/UglyToad.PdfPig/Graphics/Colors/ColorSpaces/DeviceRgbColorSpaceDetails.cs
@@ -1,0 +1,75 @@
+﻿namespace UglyToad.PdfPig.Graphics.Colors
+{
+    using System;
+
+    /// <summary>
+    /// Color values are defined by three components representing the intensities of the additive primary colorants red, green and blue.
+    /// Each component is specified by a number in the range 0.0 to 1.0, where 0.0 denotes the complete absence of a primary component and 1.0 denotes maximum intensity.
+    /// </summary>
+    public sealed class DeviceRgbColorSpaceDetails : ColorSpaceDetails
+    {
+        /// <summary>
+        /// The single instance of the <see cref="DeviceRgbColorSpaceDetails"/>.
+        /// </summary>
+        public static readonly DeviceRgbColorSpaceDetails Instance = new DeviceRgbColorSpaceDetails();
+
+        /// <inheritdoc/>
+        public override int NumberOfColorComponents => 3;
+
+        /// <inheritdoc/>
+        public override int BaseNumberOfColorComponents => NumberOfColorComponents;
+
+        private DeviceRgbColorSpaceDetails() : base(ColorSpace.DeviceRGB)
+        { }
+
+        /// <inheritdoc/>
+        internal override double[] Process(params double[] values)
+        {
+            return values;
+        }
+
+        /// <inheritdoc/>
+        public override IColor GetColor(ReadOnlySpan<double> values)
+        {
+            if (values.Length != NumberOfColorComponents)
+            {
+                throw new ArgumentException($"Invalid number of inputs, expecting {NumberOfColorComponents} but got {values.Length}", nameof(values));
+            }
+
+            double r = values[0];
+            double g = values[1];
+            double b = values[2];
+            if (r == 0 && g == 0 && b == 0)
+            {
+                return RGBColor.Black;
+            }
+
+            if (r == 1 && g == 1 && b == 1)
+            {
+                return RGBColor.White;
+            }
+
+            return new RGBColor(r, g, b);
+        }
+
+        /// <inheritdoc/>
+        public override IColor GetInitializeColor()
+        {
+            return RGBColor.Black;
+        }
+
+        /// <inheritdoc/>
+        public override void GetRgb(ReadOnlySpan<double> values, out double r, out double g, out double b)
+        {
+            r = values[0];
+            g = values[1];
+            b = values[2];
+        }
+
+        /// <inheritdoc/>
+        internal override Span<byte> Transform(Span<byte> decoded)
+        {
+            return decoded;
+        }
+    }
+}
